final-test/requirements.txt
ccxt==4.1.30
pandas==2.1.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
talib-binary==0.4.19
scipy==1.11.4
scikit-learn==1.3.2
plotly==5.17.0
jupyter==1.0.0
ipykernel==6.26.0
